requirements.txt
streamlit==1.41.1
openai==1.61.0
torch==2.6.0
torchvision==0.21.0
transformers==4.48.2
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1

